% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingMed_Abnormal_immunoglobulin.R
\name{KingMed_Abnormal_immunoglobulin}
\alias{KingMed_Abnormal_immunoglobulin}
\title{To clear immunoglobulin data from King Med test to a dataframe}
\usage{
KingMed_Abnormal_immunoglobulin(filename, WriteExecl = TRUE)
}
\arguments{
\item{filename}{the name of fusion gene data text}

\item{WriteExecl}{A logical object. TRUE as default to write excel}
}
\value{
a dataframe
}
\description{
To clear immunoglobulin data from King Med test to a dataframe
}
\examples{
KingMed_Abnormal_immunoglobulin("data.txt")
}
