% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingMed_Chromosome.R
\name{KingMed_Chromosome}
\alias{KingMed_Chromosome}
\title{To clear chromosome data from King Med test to a dataframe}
\usage{
KingMed_Chromosome(filename, WriteExecl = TRUE)
}
\arguments{
\item{filename}{the name of chromosome data text}

\item{WriteExecl}{A logical object. TRUE as default to write excel}
}
\value{
a dataframe
}
\description{
To clear chromosome data from King Med test to a dataframe
}
\examples{
KingMed_Chromosome("data.txt")
}
