% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingMed_Flow.R
\name{KingMed_Flow}
\alias{KingMed_Flow}
\title{To clear flow data from King Med test to a dataframe}
\usage{
KingMed_Flow(filename, WriteExecl = TRUE)
}
\arguments{
\item{filename}{the name of flow data text}

\item{WriteExecl}{A logical object. TRUE as default to write excel}
}
\value{
a dataframe
}
\description{
To clear flow data from King Med test to a dataframe
}
\examples{
KingMed_Flow("data.txt")
}
