% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingMed_FusionGene43.R
\name{KingMed_FusionGene43}
\alias{KingMed_FusionGene43}
\title{To clear 43 fusion genes data from King Med test to a dataframe}
\usage{
KingMed_FusionGene43(filename, WriteExecl = TRUE)
}
\arguments{
\item{filename}{the name of fusion gene data text}

\item{WriteExecl}{A logical object. TRUE as default to write excel}
}
\value{
a dataframe
}
\description{
To clear 43 fusion genes data from King Med test to a dataframe
}
\examples{
KingMed_FusionGene43("data.txt")
}
