% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingMed_WT1.R
\name{KingMed_WT1}
\alias{KingMed_WT1}
\title{To clear WT1 data from King Med test to a dataframe}
\usage{
KingMed_WT1(filename, WriteExecl = TRUE)
}
\arguments{
\item{filename}{the name of fusion gene data text}

\item{WriteExecl}{A logical object. TRUE as default to write excel}
}
\value{
a dataframe
}
\description{
To clear WT1 data from King Med test to a dataframe
}
\examples{
KingMed_WT1("data.txt")
}
